requirements.txt
Flask==2.1.0
python-dotenv==1.0.0
openai==0.27.4
gunicorn==20.1.0
